better logging

+++ card_games/automation/clear_expired_battles.py
@@ -14,6 +14,7 @@ async def clear_expired_battles(client, db, message):
     current_time = time.time()
     for battle in all_battles:
         if battle['expire_time'] < current_time:
+            print('battle with display: '+battle['card_display']+' expired')
             
             user = users.find_one({'discord_id': battle['user_id']})
             user_battle_cards = user['battle_cards']
@@ -22,7 +23,8 @@ async def clear_expired_battles(client, db, message):
                 users.update_one({'discord_id': user['discord_id']}, {'$set': {'battle_cards': user_battle_cards}})
 
             battle_message = await get_message_by_channel_and_id(client, constants.CARD_BATTLE_CHANNEL, battle['message_id'])
-            await battle_message.delete()
+            if battle_message:
+                await battle_message.delete()
 
             card_battles.delete_one({'card_display': battle['card_display']})
